refactor: extract anonymization routes into routes/anonymization.ts
Pure structural move out of createApp; no handler logic changed. Moves the
anon-control (GET/POST), anon-entities (GET/POST + suppress/unsuppress),
redacted-export, deobfuscate, and translate-query routes verbatim into a new
registerAnonymizationRoutes(app, ctx).

No new RouteContext members: the three anon stores (AnonControlStore,
CustomEntitiesStore, DiscoveredEntitiesStore) are stateless disk-backed wrappers
over ctx.store, rebuilt identically in-module (threatIntel precedent); the two
shared helpers reused (applyDeobfuscationToCase, resynthesizeInBackground) were
already graduated for the import domain. Removed now-dead server.ts imports
(AnonControl, sanitizeCustomEntities, AnonTokenCategory, deriveKnownEntities,
resolveRedactedExportOptions, redactedExportFilename, buildRedactedExport,
CustomerStore, normalizeHuntPlatform, HUNT_PLATFORMS).
